Bugfix: Use proper ButtonCodes for X11 key events
Instead of assuming that the X11 keyboard driver uses scancode + 8 as
the KeyCode, and that ButtonCodes are always equal to scancodes, use the
list of key names to create a proper mapping from keycodes used on the
local system to ButtonCodes. The X11 key names identify keys in a unique
way.

The assumption that the X11 keyboard driver always uses scancode + 8 for
the KeyCode might be correct for most modern installations, but it's too
unreliable.

The assumption that ButtonCode = scancode is incorrect for some keys,
e.g. the arrow keys.